@@ -1,24 +1,28 @@
 using Microsoft.EntityFrameworkCore;
 using comjustinspicer.Data.Models.Blog;
 
-namespace comjustinspicer.Data
+namespace comjustinspicer.Data;
+
+public class BlogContext : DbContext
 {
-	public class BlogContext : DbContext
+	public BlogContext(DbContextOptions<BlogContext> options)
+		: base(options)
 	{
-		private string _dbPath = string.Empty;
+	}
 
-		public DbSet<Post> Posts {get; set;}
+	public DbSet<Post> Posts { get; set; } = null!;
 
-		public BlogContext(){
-			var folder = Environment.SpecialFolder.LocalApplicationData;
-			var path = Environment.GetFolderPath(folder);
-			_dbPath = System.IO.Path.Join(path, "blogging.db"); //maybe this should change...
+	protected override void OnModelCreating(ModelBuilder modelBuilder)
+	{
+		base.OnModelCreating(modelBuilder);
 
-			Console.WriteLine($"Database path: {_dbPath}.");
-		}
-
-		protected override void OnConfiguring(DbContextOptionsBuilder options){
-			options.UseSqlite($"Data Source={_dbPath}");
-		}
+		modelBuilder.Entity<Post>(entity =>
+		{
+			entity.HasKey(e => e.Id);
+			entity.Property(e => e.Title)
+				.IsRequired()
+				.HasMaxLength(500);
+			entity.ToTable("Posts");
+		});
 	}
 }
