@@ -1,7 +1,8 @@
-using System;
+using System.Collections.Generic;
 
-namespace comjustinspicer.Models.Blog{
-	public class BlogViewModel {
-		public List<PostViewModel> Posts {get; set;} = new List<PostViewModel>();
-	}
+namespace comjustinspicer.Models.Blog;
+
+public sealed class BlogViewModel
+{
+	public List<PostViewModel> Posts { get; set; } = new();
 }
